+++ src/parser/block/heading.ts
@@ -9,12 +9,12 @@ import { html } from 'typed-dom';
 
 const syntax = /^(#{1,6})\s+([^\n]+)(?:\n|$)/;
 
-export const heading: HeadingParser = block(line(match(syntax, ([, { length: level }, title]) =>
+export const heading: HeadingParser = block(line(match(syntax, ([, { length: level }, content]) =>
   transform<HeadingParser>(compress(trim(some(union([indexer, inline])))), cs => {
     assert(level > 0 && level < 7);
     const el = html(`h${level}` as 'h1', cs);
     if (!hasText(el)) return;
     void defineIndex(el);
     return [[el], ''];
-  })(title)
+  })(content)
 ), true, true));